refactor: reorganize project structure by moving auth feature into feature-specific directory and add SignUpViewModel tests
