+++ app/src/main/java/must/kdroiders/hustlehub/di/AppModule.kt
@@ -10,12 +10,12 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import must.kdroiders.hustlehub.data.model.User
-import must.kdroiders.hustlehub.ui.auth.domain.repository.AuthRepository
-import must.kdroiders.hustlehub.ui.auth.data.repository.AuthRepositoryImpl
-import must.kdroiders.hustlehub.ui.auth.domain.repository.LoginResult
+import must.kdroiders.hustlehub.ui.features.auth.domain.repository.AuthRepository
+import must.kdroiders.hustlehub.ui.features.auth.data.repository.AuthRepositoryImpl
+import must.kdroiders.hustlehub.ui.features.auth.domain.repository.LoginResult
 import must.kdroiders.hustlehub.data.repository.UserRepository
 import must.kdroiders.hustlehub.data.repository.UserRepositoryImpl
-import must.kdroiders.hustlehub.ui.auth.data.remote.AuthApiService
+import must.kdroiders.hustlehub.ui.features.auth.data.remote.AuthApiService
 import must.kdroiders.hustlehub.data.remote.UserApiService
 import must.kdroiders.hustlehub.data.remote.MediaApiService
 import must.kdroiders.hustlehub.datastore.UserPreferences
