+++ app/src/main/java/must/kdroiders/hustlehub/ui/features/auth/domain/usecase/CheckUserProfileUseCase.kt
@@ -1,4 +1,4 @@
-package must.kdroiders.hustlehub.ui.auth.domain.usecase
+package must.kdroiders.hustlehub.ui.features.auth.domain.usecase
 
 import must.kdroiders.hustlehub.data.repository.UserRepository
 import javax.inject.Inject
