+++ app/src/main/java/must/kdroiders/hustlehub/ui/features/auth/domain/usecase/GetCurrentUserUseCase.kt
@@ -1,7 +1,7 @@
-package must.kdroiders.hustlehub.ui.auth.domain.usecase
+package must.kdroiders.hustlehub.ui.features.auth.domain.usecase
 
 import com.google.firebase.auth.FirebaseUser
-import must.kdroiders.hustlehub.ui.auth.domain.repository.AuthRepository
+import must.kdroiders.hustlehub.ui.features.auth.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class GetCurrentUserUseCase @Inject constructor(
