+++ app/src/main/java/must/kdroiders/hustlehub/ui/features/profilesetup/presentation/viewModel/ProfileSetupViewModel.kt
@@ -16,7 +16,7 @@ import kotlinx.coroutines.launch
 import must.kdroiders.hustlehub.data.model.User
 import must.kdroiders.hustlehub.data.model.UserRole
 import must.kdroiders.hustlehub.data.repository.UserRepository
-import must.kdroiders.hustlehub.ui.auth.domain.usecase.SyncUserProfileUseCase
+import must.kdroiders.hustlehub.ui.features.auth.domain.usecase.SyncUserProfileUseCase
 import timber.log.Timber
 import javax.inject.Inject
